requirements.txt
asgiref==3.5.2
asn1crypto==1.5.1
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
distlib==0.3.6
Django==4.1.1
django-autoslug==1.9.8
django-tinymce==3.5.0
easygui==0.98.3
endesive==2.0.12
exceptiongroup==1.1.1
filelock==3.8.0
h11==0.14.0
idna==3.4
lxml==4.9.1
msgbox==0.6.0
mysqlclient==2.1.1
numpy==1.23.5
opencv-python==4.7.0.72
oscrypto==1.3.0
outcome==1.2.0
packaging==23.1
pandas==1.5.2
paramiko==2.12.0
pdf2image==1.16.3
pdfminer==20191125
Pillow==9.2.0
platformdirs==2.5.4
poppler-utils==0.1.0
pycparser==2.21
pycryptodome==3.15.0
PyKCS11==1.5.11
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyPdf==1.13
PyPDF2==2.11.2
pyrfc==2.7.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-digital-certificate==0.3.0
python-dotenv==1.0.0
pytz==2022.6
pyusb==1.2.1
pywin32==306
pyzbar==0.1.9
reportlab==3.6.12
requests==2.30.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2022.2
urllib3==2.0.2
virtualenv==20.17.0
webdriver-manager==3.8.6
wsproto==1.2.0
